Delete user helper move methods to controllers
