+++ app/views/users/show.html.erb
@@ -26,7 +26,7 @@
   <article>
     <% if current_user.id == @user.id %>
       <ul>
-        <h3>Recived Friend Requests:</h3>
+        <h3>Received Friend Requests:</h3>
         <% current_user.friend_requests.each do |friend| %>
           <li><%= friend.name %></li>
           <%= button_to "Accept Request", {:controller => "users", :action => "confirm", :friend_id => friend.id}, :method=>:post %>
